mqtt/requirements.txt
paho-mqtt
matplotlib
numpy
dotenv
requests